+++ test/event-01.dart
@@ -11,7 +11,7 @@ void main() {
     print('trigger');
     $('div.button').trigger('click', data: 88);
   });
-  $('#trigger').click((event)=>print (event.type));
+  $('#trigger').click((QueryEvent event) { print(event.type);});
   $('#input1').change( (QueryEvent event) {
    var  value=(event.target as InputElement).value;
     print("${event.target}, value:${value}");