+++ test/event-02-blur.dart
@@ -4,20 +4,69 @@ import 'package:dquery/dquery.dart';
 
 void main() {
   
-  int docHit = 1;
-  int elemHit = 1;
-  
-  final Element docRes = querySelector('#doc-result');
-  final Element elemRes = querySelector('#elem-result');
+  //blur
+  final Element blurSec = querySelector('.blur');
   
   $document().on('blur', (QueryEvent e) {
-    docRes.innerHtml = '${docHit++}';
+    printDoc(blurSec);
+  }, selector: '.blur input');
     
-  }, selector: '#input');
-  
-  $('#input').on('blur', (QueryEvent e) {
-    elemRes.innerHtml = '${elemHit++}';
-    
+  $('.blur input').on('blur', (QueryEvent e) {
+    printElem(blurSec);
   });
   
+  
+  //focus
+  final Element focusSec = querySelector('.focus');
+  
+  $document().on('focus', (QueryEvent e) {
+    printDoc(focusSec);
+  }, selector: '.focus input');
+    
+  $('.focus input').on('focus', (QueryEvent e) {
+    printElem(focusSec);
+  });
+  
+  //click
+  final Element clickSec = querySelector('.click');
+  
+  $document().on('click', (QueryEvent e) {
+    printDoc(clickSec);
+  }, selector: '.click button');
+    
+  $('.click button').on('click', (QueryEvent e) {
+    printElem(clickSec);
+  });
+  
+  //click
+  final Element enterSec = querySelector('.mouseenter');
+  
+  $document().on('mouseenter', (QueryEvent e) {
+    printDoc(enterSec);
+  }, selector: '.mouseenter .test-area');
+    
+  $('.mouseenter .test-area').on('mouseenter', (QueryEvent e) {
+    printElem(enterSec);
+  });
+}
+
+String NUM_DOC = 'num_doc';
+String NUM_ELEM = 'num_elem';
+
+void printDoc(Element sec) {
+  int n = $(sec).data.get(NUM_DOC);
+  if (n == null)
+    n = 0;
+  
+  $(sec).data.set(NUM_DOC, ++n);
+  sec.querySelector('.doc-result').text = 'doc: $n';
+}
+
+void printElem(Element sec) {
+  int n = $(sec).data.get(NUM_ELEM);
+  if (n == null)
+    n = 0;
+  
+  $(sec).data.set(NUM_ELEM, ++n);
+  sec.querySelector('.elem-result').text = 'elem: $n';
 }